FIX BUG WITH DM

+++ index.js
@@ -64,6 +64,7 @@ var servers = {};
 
 if(message.content.startsWith(`${prefix}play`))
 {
+  if(message.type === "dm") return;
   if (!args[0]) {
          message.channel.send("Пожалуйста, предоставьте ссылку.");
          return
@@ -102,6 +103,7 @@ var servers = {};
 
 if(message.content === `${prefix}stop`)
 {
+  if(message.type === "dm") return;
   var server = servers[message.guild.id];
   if(message.guild.voiceConnection) message.guild.voiceConnection.disconnect();
 }
@@ -110,6 +112,7 @@ if(message.content === `${prefix}stop`)
 client.on("message", (message) => {
   if(message.content === `${prefix}sinfo`)
   {
+    if(message.type === "dm") return;
     let embed112 = new Discord.RichEmbed()
     .setAuthor(`Информация о сервере ${message.guild.name}`,message.guild.iconURL)
     .setColor("RANDOM")
@@ -124,7 +127,8 @@ client.on("message", (message) => {
     .addField('Эмодзи',message.guild.emojis.size,true)
     .addField('Уровень верификации',message.guild.verificationLevel,true)
     .addField('Создан:',moment(message.guild.createdAt).fromNow(),true)
-    .setThumbnail(message.guild.iconURL)
+    .setThumbnail(message.guild.iconURL);
+    if(message.type === "dm") return;
     message.channel.send(embed112)
   }
 });
@@ -155,6 +159,7 @@ client.on("message", (message) => {
 
   if(message.content.startsWith(`${prefix}userinfo`))
   {
+    if(message.type === "dm") return;
       let messageArray = message.content.split(" ");
       let args = messageArray.slice(1);
 
@@ -210,6 +215,7 @@ client.on("message", (message) => {
 
   if(message.content.startsWith(`${prefix}prune`))
   {
+    if(message.type === "dm") return;
     if(!message.member.hasPermission("MANAGE_MESSAGES")) return message.reply("отсутствуют права");
     if(!args[0]) return message.reply("нельзя очистить меньше 1 знака или 0");
     message.channel.bulkDelete(args[0]).then(() => {
@@ -221,6 +227,7 @@ client.on("message", (message) => {
 client.on("message", (message) => {
   if(message.content.startsWith(`${prefix}mute`))
   {
+    if(message.type === "dm") return;
     let mute = message.guild.roles.find(role => role.name === "AlfredMute");
     let m = message.guild.member(message.mentions.users.first() || message.guild.members.get(args[0]));
     if(message.member.hasPermissions("MANAGE_MESSAGES"))
@@ -251,6 +258,7 @@ client.on('message',async(message) =>{
   let args = messageArray.slice(1);
 if(message.content.startsWith(`${prefix}poll`))
   {
+    if(message.type === "dm") return;
   message.delete();
     if (!args) return message.reply("Вы должны указать вопрос голосования!")
     if (!message.content.includes("?")) return message.reply("Добавьте `?` в конце,чтобы начать голосование")
@@ -271,6 +279,7 @@ client.on("message", (message) => {
 
   if(message.content.startsWith(`${prefix}ban`))
   {
+    if(message.type === "dm") return;
     let bu = message.guild.member(message.mentions.users.first() || message.guild.members.get(args[0]));
 
     if(!bu) return message.channel.send(`Пользователь не найден.`);
@@ -304,6 +313,7 @@ client.on("message", (message) => {
 
   if(message.content.startsWith(`${prefix}kick`))
   {
+    if(message.type === "dm") return;
     let ku = message.guild.member(message.mentions.users.first() || message.guild.members.get(args[0]));
 
     if(!ku) return message.channel.send(`Пользователь не найден.`);
@@ -389,6 +399,7 @@ client.on("message", (message) => {
 client.on("message", (message) => {
   if(message.content.startsWith(`${prefix}avatar`))
   {
+    if(message.type === "dm") return;
      let user = message.mentions.users.first() || message.author;
     const embed = new Discord.RichEmbed()
     .setColor('RANDOM')
